backend/playto-community-feed/requirements.txt
asgiref==3.11.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.0.0
packaging==26.0
sqlparse==0.5.5
tzdata==2025.3
